Change calls to .IsNothing() to .Check(), which seems more appropriate.
Swig maintainers make a good case for leaving the warning-messages intact, as this hack prevents anyone from testing the return-value of a Set() operation.

I've no problem with that - they have to answer to a wider user base. This here is a personal fork for *MY* purposes, refactoring a poorly-maintained library
so that I can eventually use it with TypeScript (NodeJS) on a current revision of Node, within a current revision of Alpine, within a Containerized setting.